requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
plotly>=5.17.0
matplotlib>=3.7.0
pandas>=2.0.0